edit order

+++ app/Http/Controllers/Admin/RoleController.php
@@ -5,12 +5,14 @@ namespace App\Http\Controllers\Admin;
 //use App\Models\Role;
 //use App\Models\Permission;
 use Illuminate\Http\Request;
-use App\Http\Controllers\Controller;
-use App\Http\Requests\UpdateRoleRequest;
-use App\Contracts\RoleRepositoryInterface;
 use Spatie\Permission\Models\Role;
-use Spatie\Permission\Models\Permission;
+use Illuminate\Support\Facades\Log;
+use App\Http\Controllers\Controller;
 use Spatie\Permission\Traits\HasRoles;
+use App\Http\Requests\StoreRoleRequest;
+use App\Http\Requests\UpdateRoleRequest;
+use Spatie\Permission\Models\Permission;
+use App\Contracts\RoleRepositoryInterface;
 
 class RoleController extends Controller
 {
@@ -26,18 +28,21 @@ class RoleController extends Controller
     }
     public function index(){
         $roles=$this->roleRepository->getAllRolesWithPermissions();
-        return view('admin.roles.all-role',['roles' => $roles]);     }
+        return view('admin.roles.all-role',['roles' => $roles]);     
+    }
     public function create(){
         $permissions = $this->roleRepository->getAllPermissions();
         return view('admin.roles.create-role')->with('permissions',$permissions);
     }
-    public function store(UpdateRoleRequest $request){
+    public function store(StoreRoleRequest $request){
         $validatedData=$request->validated();    
-        $role= app(RoleRepositoryInterface::class)->createRole($validatedData);
-        $role->permissions()->sync($validatedData['permissions']);
+        $role= app(RoleRepositoryInterface::class)->store($validatedData);
+        $permissions = Permission::whereIn('id', $request->permissions)->get(['name'])->toArray();
+        $role->syncPermissions($permissions);
         return redirect(route('roles.index'))->with('message','Data saved.');
     }
     public function edit($id){
+
         $permissions = $this->roleRepository->getAllPermissions();
         $rolePermissions=$this->roleRepository->getRolePermissions($id);
         $role=$this->roleRepository->findRoleById($id);
@@ -47,24 +52,31 @@ class RoleController extends Controller
             'rolePermissions' => $rolePermissions]);
     }
     public function update(UpdateRoleRequest $request,$id){
-       
-        $validatedName= $request->only('name');
-        $role=app(RoleRepositoryInterface::class)->updateRole($validatedName,$id);
-        $validatedData = $request->validated();
-        $role->permissions()->sync($validatedData['permissions']);
+
+       $validatedName= $request->only('name');
+       $role=app(RoleRepositoryInterface::class)->update($validatedName,$id);
+
+       $validatedData = $request->validated();
+       $permissions = Permission::whereIn('id', $request->permissions)->get(['name'])->toArray();
+       $role->syncPermissions($permissions);
+
         return redirect(route('roles.index'))->with('message','Data updated.');
     }
     
-    public function destroy(Role $role){
+    public function destroy($id){
+
+        $role =  $this->roleRepository->findRoleById($id);
+
         if($role->name=='Super Admin'){
             abort(403, 'SUPER ADMIN ROLE CAN NOT BE DELETED');
         }
-        // if(auth()->user()->hasRole($role->name)){
-        //     abort(403, 'CAN NOT DELETE SELF ASSIGNED ROLE');
-        // }
-        $role= $this->roleRepository->destroy($role->id);
+        if(auth()->user()->hasRole($role->name)){
+            abort(403, 'CAN NOT DELETE SELF ASSIGNED ROLE');
+        }
+    
+        $role= $this->roleRepository->destroy($id);
         $role->delete();
-        return redirect()->route('roles.index')
-                ->withSuccess('Role is deleted successfully.');
+        return response()->json([ 'status' => 'role deleted successfully' ]);
+
     }
 }
